Removed publishing the artifact to Maven [#4]
If we try to publish this to the Maven repository so that ComiXed
can pull it down as part of a release, then it becomes a chicken
and egg thing of which release to publish first.

Instead, CX users are going to need to download the releases for
themselves and install them.